moved includes from .h to .c where needed

+++ include/read_sequencev2.h
@@ -3,17 +3,18 @@
 
 #include <complex.h>
 #include <stddef.h>
+#include <stddef.h>
 
 int read_sequence(const char *filename, double complex **data_out,
-                  size_t *len_out, const int MAX_INPUT);
+                  size_t *len_out, const size_t MAX_INPUT);
 
 int read_complex_text(const char *filename, double complex **data_out,
-                      size_t *len_out, const int MAX_INPUT);
+                      size_t *len_out, const size_t MAX_INPUT);
 
 int read_complex_csv(const char *filename, double complex **data_out,
-                     size_t *len_out, const int MAX_INPUT);
+                     size_t *len_out, const size_t MAX_INPUT);
 
 int read_complex_bin(const char *filename, double complex **data_out,
-                     size_t *len_out, const int MAX_INPUT);
+                     size_t *len_out, const size_t MAX_INPUT);
 
 #endif
